Fix Windows build Change bone index vertex attribute from unsigned int to float, since AMD Win32 drivers don't seem to like addressing uniform arrays with uvec4 attributes for some reason
git-svn-id: http://ludobloom.com/svn/StemLibProjects@2653 d291a34c-4c4d-0410-8556-a73de5c6a500

+++ glgraphics/trunk/resources/animated.vert
@@ -5,7 +5,7 @@ in vec2 inTexCoord;
 in vec3 inNormal;
 in vec4 inTangent;
 in vec4 inColor;
-in uvec4 inBoneID;
+in vec4 inBoneID;
 in vec4 inBoneWeight;
 
 #define BONE_COUNT_MAX 128
@@ -31,10 +31,10 @@ void main() {
 	texCoord = inTexCoord;
 	color = inColor * materialColor;
 	
-	weightedNormals[0] = mat3(boneTransforms[inBoneID.x]) * inNormal * inBoneWeight.x;
-	weightedNormals[1] = mat3(boneTransforms[inBoneID.y]) * inNormal * inBoneWeight.x;
-	weightedNormals[2] = mat3(boneTransforms[inBoneID.z]) * inNormal * inBoneWeight.x;
-	weightedNormals[3] = mat3(boneTransforms[inBoneID.w]) * inNormal * inBoneWeight.x;
+	weightedNormals[0] = mat3(boneTransforms[int(inBoneID.x)]) * inNormal * inBoneWeight.x;
+	weightedNormals[1] = mat3(boneTransforms[int(inBoneID.y)]) * inNormal * inBoneWeight.x;
+	weightedNormals[2] = mat3(boneTransforms[int(inBoneID.z)]) * inNormal * inBoneWeight.x;
+	weightedNormals[3] = mat3(boneTransforms[int(inBoneID.w)]) * inNormal * inBoneWeight.x;
 	normal = weightedNormals[0] + weightedNormals[1] + weightedNormals[2] + weightedNormals[3];
 	
 	normalMatrix = mat3(transpose(inverse(modelTransform))); // TODO: Calculate this outside the shader
@@ -43,10 +43,10 @@ void main() {
 	bitangent = inTangent.w * cross(normal, tangent);
 	tbnMatrix = mat3(tangent, bitangent, normal);
 	
-	weightedPositions[0] = boneTransforms[inBoneID.x] * vec4(inPosition, 1.0) * inBoneWeight.x;
-	weightedPositions[1] = boneTransforms[inBoneID.y] * vec4(inPosition, 1.0) * inBoneWeight.y;
-	weightedPositions[2] = boneTransforms[inBoneID.z] * vec4(inPosition, 1.0) * inBoneWeight.z;
-	weightedPositions[3] = boneTransforms[inBoneID.w] * vec4(inPosition, 1.0) * inBoneWeight.w;
+	weightedPositions[0] = boneTransforms[int(inBoneID.x)] * vec4(inPosition, 1.0) * inBoneWeight.x;
+	weightedPositions[1] = boneTransforms[int(inBoneID.y)] * vec4(inPosition, 1.0) * inBoneWeight.y;
+	weightedPositions[2] = boneTransforms[int(inBoneID.z)] * vec4(inPosition, 1.0) * inBoneWeight.z;
+	weightedPositions[3] = boneTransforms[int(inBoneID.w)] * vec4(inPosition, 1.0) * inBoneWeight.w;
 	weightedPosition = weightedPositions[0] + weightedPositions[1] + weightedPositions[2] + weightedPositions[3];
 	position = vec3(modelTransform * weightedPosition);
